backend/requirements.txt
alembic==1.16.5
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
backports.asyncio.runner==1.2.0
bcrypt==5.0.0
black==25.9.0
certifi==2025.10.5
click==8.1.8
coverage==7.10.7
exceptiongroup==1.3.0
fastapi==0.120.0
flake8==7.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.1.0
Mako==1.3.10
MarkupSafe==3.0.3
mccabe==0.7.0
mypy==1.18.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pydantic==2.12.3
pydantic_core==2.41.4
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-asyncio==1.2.0
pytest-cov==7.0.0
python-dotenv==1.2.1
pytokens==0.2.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
tomli==2.3.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
